test(sidecar-ffi): add file for span tests
